+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - Added About page
 - Added blog component
 - Added blogs page and component
+- Added custom Carousel